requirements.txt
pyTelegramBotAPI==4.14.0
Flask==3.0.3